requirements.txt
# Core packages
crewai>=0.11.0
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.7.1

# Additional dependencies
python-dotenv>=1.0.0
tiktoken>=0.5.2
pydantic>=2.7.1
crewai-tools>=0.0.14
scholarly>=1.7.11
PyPDF2>=3.0.0

# Optional packages
ipykernel>=6.29.4
ipython>=8.24.0
jupyter_client>=8.6.2
jupyter_core>=5.7.2
langchain-community>=0.0.38
langchain-core>=0.1.52
langchain-text-splitters>=0.0.2
langsmith>=0.1.63
yfinance>=0.2.40
